Arrumei as barras para considerar o aumento de dano quando os jogadores estão quase morrendo

+++ funcoes.py
@@ -77,12 +77,21 @@ def tecla_pressionada(player1, player2, segundos):
     return True
 
 def encostou(player1, player2):
+    from game_screen import rodada
     if player2.image in player2.punch_img or player2.image in player2.goldpunch_img:
         if player2.umsoco == 1:
             player1.life -= player2.damage
             player2.umsoco = 0
-            if player1.compbarraverd > 0:
-                player1.compbarraverd -= 20
+            if player2.image in player2.punch_img:
+                if player1.compbarraverd > 0 and rodada == 1:
+                    player1.compbarraverd -= 30
+                elif player1.compbarraverm > 0 and rodada == 2:
+                    player1.compbarraverm -= 30
+            elif player2.image in player2.goldpunch_img:
+                if player1.compbarraverd > 0 and rodada == 1:
+                    player1.compbarraverd -= 45
+                elif player1.compbarraverm > 0 and rodada == 2:
+                    player1.compbarraverm -= 45
     if player1.rect.centerx >= player2.rect.centerx:
         player1.rect.x += 30
     if player1.rect.centerx < player2.rect.centerx:
@@ -93,8 +102,16 @@ def encostou(player1, player2):
         if player1.umsoco == 1:
             player2.life -= player1.damage
             player1.umsoco = 0
-            if player2.compbarraverd > 0:
-                player2.compbarraverd -= 20
+            if player1.image in player1.punch_img:
+                if player2.compbarraverd > 0 and rodada == 1:
+                    player2.compbarraverd -= 30
+                elif player2.compbarraverm > 0 and rodada == 2:
+                    player2.compbarraverm -= 30
+            elif player1.image in player1.goldpunch_img:
+                if player2.compbarraverd > 0 and rodada == 1:
+                    player2.compbarraverd -= 45
+                elif player2.compbarraverm > 0 and rodada == 2:
+                    player2.compbarraverm -= 45
     if player2.rect.centerx >= player1.rect.centerx:
         player2.rect.x += 30
     if player2.rect.centerx < player1.rect.centerx:
@@ -104,14 +121,17 @@ def encostou(player1, player2):
     return 0
 
 def verificamorreu(player1, player2):
-    if player1.life<=0:     
+    from game_screen import rodada
+    if player1.life<=0 and rodada <= 3:     
         player1.compbarraverd = 0       
         player1.kill()
         player1.life = 0
-    if player2.life<=0:
+        rodada += 1
+    if player2.life<=0 and rodada <= 3:
         player2.compbarraverd = 0
         player2.kill()
         player2.life = 0
+        rodada += 1
     return 0
 
 def contadorinicial(font,pos,all_sprites):
